requirements.txt
streamlit
torch
transformers
pandas
Pillow
chromadb
